move Go types to different files
