+++ golang/task.go
@@ -1,45 +1,8 @@
 package main
 
 import (
-	"fmt"
-	"strconv"
 	"time"
 )
-
-var timeFormat = time.DateOnly
-
-type deadline struct {
-	date time.Time
-}
-
-func NewDeadline(deadlineString string) (deadline, error) {
-	date, err := time.Parse(timeFormat, deadlineString)
-	if err != nil {
-		return deadline{}, err
-	}
-
-	return deadline{
-		date: date,
-	}, nil
-}
-
-func (d *deadline) String() string {
-	return fmt.Sprintf(" (%s)", d.date.Format(timeFormat))
-}
-
-func (d *deadline) IsEmpty() bool {
-	return d.date.IsZero()
-}
-
-type identifier int64
-
-func NewIdentifier(idString string) (identifier, error) {
-	id, err := strconv.ParseInt(idString, 10, 64)
-	if err != nil {
-		return -1, err
-	}
-	return identifier(id), nil
-}
 
 // Task describes an elementary task.
 type Task struct {
